Use monaco-worker-manager
I extracted this logic from monaco-yaml, so it can be reused.
