+++ src/index.ts
@@ -1,25 +1,38 @@
 import { languages } from 'monaco-editor/esm/vs/editor/editor.api.js';
+import { MonacoTailwindcssOptions } from 'monaco-tailwindcss';
+import { createWorkerManager } from 'monaco-worker-manager';
 
 import {
   createColorProvider,
   createCompletionItemProvider,
   createHoverProvider,
 } from './languageFeatures';
-import { createWorkerManager } from './workerManager';
+import { TailwindcssWorker } from './tailwindcss.worker';
 
 export const defaultLanguageSelector = ['javascript', 'html', 'mdx', 'typescript'] as const;
 
 export const configureMonacoTailwindcss: typeof import('monaco-tailwindcss').configureMonacoTailwindcss =
   ({ config, languageSelector = defaultLanguageSelector } = {}) => {
-    const getWorker = createWorkerManager({ config });
+    const workerManager = createWorkerManager<TailwindcssWorker, MonacoTailwindcssOptions>({
+      label: 'tailwindcss',
+      moduleId: 'monaco-tailwindcss/tailwindcss.worker',
+      createData: { config },
+    });
 
     const disposables = [
-      languages.registerColorProvider(languageSelector, createColorProvider(getWorker)),
+      workerManager,
+      languages.registerColorProvider(
+        languageSelector,
+        createColorProvider(workerManager.getWorker),
+      ),
       languages.registerCompletionItemProvider(
         languageSelector,
-        createCompletionItemProvider(getWorker),
+        createCompletionItemProvider(workerManager.getWorker),
       ),
-      languages.registerHoverProvider(languageSelector, createHoverProvider(getWorker)),
+      languages.registerHoverProvider(
+        languageSelector,
+        createHoverProvider(workerManager.getWorker),
+      ),
     ];
 
     return {
